+++ riot/utils/errors.py
