+++ riot/utils/objects.py
