+++ riot/utils/platform_and_region.py
